requirements.txt
python-dotenv
argparse
playwright
langchain-openai
pydantic
browser-use
chromium
langchain_community